BUG: Fixes a bug decompiling nested dict literals.
We incorrectly assumed that the first stack_builder supplied to our
dict-literal decompilation routine would always be a BUILD_MAP.  This
isn't true when a dict is nested in a complex expression.

Moreover, we incorrectly assumed that the furthest-away BUILD_MAP was
always the consumer of a STORE_MAP instruction, which wasn't true when
dict literals are complexly nested.  We now implement a generic
algorithm for finding the correct BUILD_MAP by walking backward from the
STORE_MAP being consumed, maintaining a count of the net number of
stores consumed, and breaking when that number falls below 0.