ENH: adds general purpose context
Adds CodeTransformer.context for users to dump contextual data while
performing a transformation. This allows users to persist information
across dispatched transformation functions.

A transformation context exists for the scope of transforming a single
code object. This is reset when entering the transformation of code
object found in the consts.

By default the context will contain code object being worked on and the
startcode.

The context uses thread local storage so that a single transformer may
be used in multuiple threads at the same time.